move files in new namespaces
