+++ deptrac.yaml
@@ -29,6 +29,10 @@ deptrac:
       collectors:
         - type: directory
           value: src/Cryptography/.*
+    - name: DecisionModel
+      collectors:
+        - type: directory
+          value: src/DecisionModel/.*
     - name: EventBus
       collectors:
         - type: directory
@@ -72,6 +76,10 @@ deptrac:
               value: MetadataMessage
             - type: layer
               value: MetadataSubscriber
+    - name: Projection
+      collectors:
+        - type: directory
+          value: src/Projection/.*
     - name: QueryBus
       collectors:
         - type: directory
@@ -100,11 +108,6 @@ deptrac:
       collectors:
         - type: directory
           value: src/Subscription/.*
-    - name: Test
-      collectors:
-        - type: directory
-          value: src/Test/.*
-
   ruleset:
     Aggregate:
       - Attribute
@@ -118,6 +121,7 @@ deptrac:
       - Identifier
       - MetadataAggregate
       - Repository
+      - Store
     Console:
       - Aggregate
       - Message
@@ -130,11 +134,12 @@ deptrac:
       - Subscription
     Cryptography:
       - Schema
-    Debug:
+    DecisionModel:
       - Attribute
       - Message
-      - Repository
-      - Subscription
+      - Projection
+      - Serializer
+      - Store
     EventBus:
       - Attribute
       - Message
@@ -161,16 +166,9 @@ deptrac:
       - Attribute
       - Metadata
       - Subscription
-    Subscription:
-      - Aggregate
+    Projection:
       - Attribute
-      - Clock
-      - Identifier
       - Message
-      - MetadataEvent
-      - MetadataSubscriber
-      - Repository
-      - Schema
       - Store
     QueryBus:
       - Attribute
@@ -182,10 +180,12 @@ deptrac:
       - MetadataAggregate
       - MetadataEvent
       - EventBus
+      - Serializer
       - Snapshot
       - Store
     Schema:
     Serializer:
+      - Attribute
       - Cryptography
       - Identifier
       - MetadataEvent
@@ -200,6 +200,17 @@ deptrac:
       - Clock
       - Message
       - Metadata
+      - MetadataEvent
       - Schema
       - Serializer
-    Test:
+    Subscription:
+      - Aggregate
+      - Attribute
+      - Clock
+      - Identifier
+      - Message
+      - MetadataEvent
+      - MetadataSubscriber
+      - Repository
+      - Schema
+      - Store
