+++ tests/Integration/DynamicConsistencyBoundary/DynamicConsistencyBoundaryTest.php
@@ -7,8 +7,8 @@ namespace Patchlevel\EventSourcing\Tests\Integration\DynamicConsistencyBoundary;
 use Doctrine\DBAL\Connection;
 use Patchlevel\EventSourcing\CommandBus\ServiceHandlerProvider;
 use Patchlevel\EventSourcing\CommandBus\SyncCommandBus;
-use Patchlevel\EventSourcing\DCB\StoreDecisionModelBuilder;
-use Patchlevel\EventSourcing\DCB\StoreEventAppender;
+use Patchlevel\EventSourcing\DecisionModel\StoreDecisionModelBuilder;
+use Patchlevel\EventSourcing\DecisionModel\StoreEventAppender;
 use Patchlevel\EventSourcing\Message\Message;
 use Patchlevel\EventSourcing\Metadata\Event\AttributeEventRegistryFactory;
 use Patchlevel\EventSourcing\Schema\DoctrineSchemaDirector;
