+++ tests/Unit/Serializer/AttributeEventTagExtractorTest.php
@@ -2,11 +2,11 @@
 
 declare(strict_types=1);
 
-namespace Patchlevel\EventSourcing\Tests\Unit\DCB;
+namespace Patchlevel\EventSourcing\Tests\Unit\Serializer;
 
 use Patchlevel\EventSourcing\Attribute\EventTag;
-use Patchlevel\EventSourcing\DCB\AttributeEventTagExtractor;
 use Patchlevel\EventSourcing\Identifier\CustomId;
+use Patchlevel\EventSourcing\Serializer\AttributeEventTagExtractor;
 use PHPUnit\Framework\TestCase;
 
 final class AttributeEventTagExtractorTest extends TestCase
